DEVELOP: Refactoring

+++ src/subcommand/install.rs
@@ -14,6 +14,8 @@ pub enum InstallError<'a> {
     MalformedPackage,
     PackageTypeIsNotImplemented,
     IOError(io::Error),
+    // This error never occurs.
+    Eden,
 }
 
 struct CmdOption {
@@ -35,16 +37,14 @@ fn parse_options<'a>(matches: &'a ArgMatches<'a>) -> Result<CmdOption, InstallEr
     Ok(opts)
 }
 
-fn clone_git_repo<'a>(
-    mimium_dir: PathBuf,
-    host: String,
-    path: String,
-) -> Result<String, InstallError<'a>> {
-    let repo_url = format!("https://{}/{}.git", host, path);
+fn clone_git_repo<'a>(mimium_dir: PathBuf, pkg: Package) -> Result<String, InstallError<'a>> {
+    let repo_url = pkg.remote_url().unwrap();
     // TODO: Protection from the directory traversal attack?
-    let path_buf = PathBuf::from(path.clone());
-    let repo_name = path_buf.file_name().unwrap().to_str().unwrap();
-    let pkg_path = format!("{}/{}/{}", mimium_dir.to_str().unwrap(), host, &repo_name);
+    let pkg_path = format!(
+        "{}/{}",
+        mimium_dir.to_str().unwrap(),
+        pkg.path().to_str().unwrap()
+    );
 
     info!("Cloning into {:?}...", pkg_path);
     // NOTE: libgit2 cannot shallow clone...
@@ -63,35 +63,37 @@ fn clone_git_repo<'a>(
     }
 }
 
-fn install_git_repo<'a>(
-    mimium_dir: PathBuf,
-    host: String,
-    path: String,
-) -> Result<(), InstallError<'a>> {
-    info!("install {:?} from {:?} as Git repository.", path, host);
-    match clone_git_repo(mimium_dir, host, path.clone()) {
-        Ok(pkg_path) => {
-            let pkg_path = PathBuf::from(pkg_path);
-            match is_mimium_package(&pkg_path) {
-                Ok(true) => Ok(()),
-                Ok(false) => {
-                    error!("The repository {} is not a mimium package.", path.clone());
-                    info!("Removing the repository {}...", path.clone());
-                    if let Err(err) = fs::remove_dir_all(pkg_path.clone()) {
-                        error!("Cannot remove repository because {}", err);
+fn install_git_repo<'a>(mimium_dir: PathBuf, pkg: Package) -> Result<(), InstallError<'a>> {
+    if let Package::Git { host, path } = pkg.clone() {
+        info!("Install {:?} from {:?} as Git repository.", path, host);
+
+        match clone_git_repo(mimium_dir, pkg) {
+            Ok(pkg_path) => {
+                let pkg_path = PathBuf::from(pkg_path);
+                match is_mimium_package(&pkg_path) {
+                    Ok(true) => Ok(()),
+                    Ok(false) => {
+                        error!("The repository {} is not a mimium package.", path.clone());
+
+                        info!("Removing the repository {}...", path.clone());
+                        if let Err(err) = fs::remove_dir_all(pkg_path.clone()) {
+                            error!("Cannot remove repository because {}", err);
+                        }
+                        Err(InstallError::MalformedPackage)
                     }
-                    Err(InstallError::MalformedPackage)
+                    Err(err) => Err(InstallError::IOError(err)),
                 }
-                Err(err) => Err(InstallError::IOError(err)),
             }
+            Err(err) => Err(err),
         }
-        Err(err) => Err(err),
+    } else {
+        Err(InstallError::Eden)
     }
 }
 
 fn install_package<'a>(mimium_dir: PathBuf, opt: CmdOption) -> Result<(), InstallError<'a>> {
-    match opt.package {
-        Package::Git { host, path } => install_git_repo(mimium_dir, host, path),
+    match opt.package.clone() {
+        Package::Git { host: _, path: _ } => install_git_repo(mimium_dir, opt.package),
         Package::Pkg(_name) => Err(InstallError::PackageTypeIsNotImplemented),
         Package::Path(_path) => Err(InstallError::PackageTypeIsNotImplemented),
     }
